calculator, header, betslip components
Split app into three components
