+++ src/components/betslip.jsx
@@ -6,26 +6,31 @@ const Betslip = ({ betslip, runners }) => {
   const [result, setResult] = useState(null);
 
   const calculateResult = () => {
-    console.log('betslip:', betslip);
-    if (betslip && typeof betslip === 'object' && betslip !== null && 'stake' in betslip && 'type' in betslip && 'eachWay' in betslip) {
+    console.log("betslip:", betslip);
+    if (betslip && typeof betslip === "object" && betslip !== null && "stake" in betslip && "type" in betslip && "eachWay" in betslip) {
       setResult(betcruncher.calculator(betslip, runners));
     }
   };
-  
+
   return (
-    <Paper sx={{ p: 2, position: "sticky", top: 10, minHeight: 250}} elevation={8}>
-      <Typography variant="h4" sx={{ mb: 2 }}>
-        {result ? `Total Stake: ${result.totalStake}` : 'Total Stake: 0'}
-      </Typography>
-      <Typography variant="h4" sx={{ mb: 2 }}>
-        {result ? `Returns: ${result.returns}` : 'Returns: 0'}
-      </Typography>
-      <Typography variant="h4" sx={{ mb: 2 }}>
-        {result ? `Profit: ${result.profit}` : 'Profit: 0'}
-      </Typography>
+    <Paper sx={{ p: 2, minHeight: 250 }} elevation={8}>
+
+      <Box sx={{ display: "flex", flexDirection: "row", justifyContent: "space-evenly", alignItems: "center" }}>
+        <Typography variant="h2" sx={{ mb: 2 }}>
+          {result ? `Total Stake: £${result.totalStake}` : "Total Stake: £0"}
+        </Typography>
+        <Typography variant="h2" sx={{ mb: 2 }}>
+          {result ? `Returns: £${result.returns}` : "Returns: £0"}
+        </Typography>
+        <Typography variant="h2" sx={{ mb: 2 }}>
+          {result ? `Profit: £${result.profit}` : "Profit: £0"}
+        </Typography>
+      </Box>
+
       <Button variant="outlined" onClick={calculateResult} fullWidth>
         Calculate
       </Button>
+    
     </Paper>
   );
 };
